requirements.txt
azure-mgmt-datafactory
azure-identity
azure-storage-file-datalake